onnxruntime并行推理/art_training/requirements.txt
certifi @ file:///croot/certifi_1671487769961/work/certifi
coloredlogs==15.0.1
decorator==5.1.1
flatbuffers==23.3.3
humanfriendly==10.0
mpmath==1.3.0
numpy==1.24.2
onnxruntime-gpu==1.13.1
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
pika==1.3.1
Pillow==9.4.0
protobuf==3.20.1
py==1.11.0
retry==0.9.2
sympy==1.11.1
typing==3.7.4.3
typing_extensions==4.5.0
